refactor(web): extract useChatMessageSync from useChatWithSync
Move the message fetch/merge family out of the 833-line useChatWithSync
hook into a focused useChatMessageSync hook: the load-on-select effect
(+ its messagesLoadFailed ref), reloadMessages, refetchMessages (delta
sync), reloadChat (post-disconnect) and addMessageToChat.

useChatWithSync composes it via a single destructure and its public
return shape is unchanged. Behavior-preserving: logic moved verbatim,
effect ordering and dependency arrays kept. useChatWithSync 833 -> 724.
